refactor(CHASE-174): 翻訳ルート定義を別ファイルに分離
activities/index.tsから翻訳関連ルート定義を
translations-body.tsに分離し、routes.tsでマウント。
関心の分離によりファイルの責務が明確化。

[claude-opus-4-20250514]
